+++ backend/src/main/java/controller/NotificationController.java
@@ -1,9 +1,12 @@
 package controller;
 
 import entity.Notification;
+import entity.Tag;
+import io.quarkus.security.UnauthorizedException;
 import jakarta.inject.Inject;
 import miscellaneous.ServiceException;
 import org.eclipse.microprofile.graphql.*;
+import org.eclipse.microprofile.jwt.JsonWebToken;
 import org.jboss.logging.Logger;
 import service.NotificationService;
 
@@ -13,6 +16,8 @@ import java.util.List;
 public class NotificationController {
 
     @Inject
+    JsonWebToken jwt;
+    @Inject
     NotificationService notificationService;
 
     private static final Logger LOG = Logger.getLogger(TagController.class.getName());
@@ -21,6 +26,10 @@ public class NotificationController {
     @Description("Fetches a list of all notifications for a user in the database")
     public List<Notification> getAllNotificationsForUserWithId(Long id) throws GraphQLException {
         LOG.info("getAllNotificationsForUserWithId");
+        // check if sender is allowed
+        if (!id.equals(Long.parseLong(jwt.getClaim("userid").toString())) && !jwt.getGroups().contains("Administrator")){
+            throw new UnauthorizedException();
+        }
         return notificationService.getAllNotificationsForUserWithId(id);
     }
 
@@ -28,6 +37,10 @@ public class NotificationController {
     @Description("Removes the given user from the given Notification and deletes it, if there are not associated users left")
     public void deleteUserFromNotification(Long userId, Long notificationId) throws GraphQLException{
         LOG.info("deleteUserFromNotification");
+        // check if sender is allowed
+        if (!userId.equals(Long.parseLong(jwt.getClaim("userid").toString())) && !jwt.getGroups().contains("Administrator")){
+            throw new UnauthorizedException();
+        }
         try{
             this.notificationService.deleteUserFromNotification(userId, notificationId);
         }catch (ServiceException e) {
